requirements.txt
fastapi
uvicorn
psycopg2-binary
bcrypt
requests
gunicorn